+++ packages/types/src/Imported.ts
@@ -5,7 +5,13 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
-export type { HexString } from '@polkadot/util/types'
-export type { AnyNumber, AnyJson } from '@polkadot/types/types/codec'
+export type {
+  ISubmittableResult,
+  AnyNumber,
+  AnyJson,
+} from '@polkadot/types/types'
 export type { BN } from '@polkadot/util'
+export type { HexString } from '@polkadot/util/types'
 export type { Prefix } from '@polkadot/util-crypto/address/types'
+export type { SubmittableExtrinsic } from '@polkadot/api/promise/types'
+export type { KeyringPair } from '@polkadot/keyring/types'
