+++ packages/types/src/index.ts
@@ -9,10 +9,6 @@
  * @module @kiltprotocol/types
  */
 
-export type { ISubmittableResult } from '@polkadot/types/types'
-export type { SubmittableExtrinsic } from '@polkadot/api/promise/types'
-export type { KeyringPair } from '@polkadot/keyring/types'
-
 export * as SubscriptionPromise from './SubscriptionPromise.js'
 
 export * from './Attestation.js'
